+++ modules/editor-one/assets/js/sidebar-navigation/components/header/styled-components.js
@@ -8,7 +8,6 @@ export const HeaderContainer = styled( Box )( ( { theme } ) => ( {
 	borderBottom: `1px solid ${ theme.palette.divider }`,
 	display: 'flex',
 	alignItems: 'center',
-	marginBottom: theme.spacing( 1 ),
 } ) );
 
 export const HeaderContent = styled( Box )( ( { theme } ) => ( {
